remove redundant cast

+++ main.c
@@ -149,7 +149,7 @@ FileList *main_file_list_from_args(int argc, char **argv) {
         file = &(file_list->files[length]);
         file->name = util_malloc(name_length+1);
         memcpy(file->name, name, name_length+1);
-        file->length = (uint32) name_length;
+        file->length = name_length;
 
         length += 1;
     }
@@ -193,7 +193,7 @@ FileList *main_file_list_from_dir(char *directory) {
         free(directory_list[i]);
         length += 1;
     }
-    file_list->length = (uint32) length;
+    file_list->length = length;
     free(directory_list);
     return file_list;
 }
